[Decomp][BattleGround] Cohesion split: extract mgr utilities to BattleGroundMgrUtil.cpp
