+++ src/game/BattleGround/BattleGroundMgr.cpp
@@ -1097,427 +1097,17 @@ void BattleGroundMgr::SendToBattleGround(Player* pl, uint32 instanceId, BattleGr
     }
 }
 
-bool BattleGroundMgr::IsArenaType(BattleGroundTypeId bgTypeId)
-{
-    switch (bgTypeId)
-    {
-        case BATTLEGROUND_NA:
-        case BATTLEGROUND_BE:
-        case BATTLEGROUND_RL:
-        case BATTLEGROUND_AA:
-            return true;
-        default:
-            return false;
-    };
-}
 
-/**
- * @brief Converts a battleground type ID to a queue type ID.
- *
- * Maps a battleground type ID to its corresponding queue type ID. Different queue types
- * have separate queues in the matchmaking system.
- *
- * @param bgTypeId The battleground type ID.
- * @param arenaType
- * @return The corresponding queue type ID, or BATTLEGROUND_QUEUE_NONE if invalid.
- */
-BattleGroundQueueTypeId BattleGroundMgr::BGQueueTypeId(BattleGroundTypeId bgTypeId, ArenaType arenaType)
-{
-    switch (bgTypeId)
-    {
-        case BATTLEGROUND_WS:
-            return BATTLEGROUND_QUEUE_WS;
-        case BATTLEGROUND_AB:
-            return BATTLEGROUND_QUEUE_AB;
-        case BATTLEGROUND_AV:
-            return BATTLEGROUND_QUEUE_AV;
-        case BATTLEGROUND_EY:
-            return BATTLEGROUND_QUEUE_EY;
-        case BATTLEGROUND_AA:
-        case BATTLEGROUND_NA:
-        case BATTLEGROUND_RL:
-        case BATTLEGROUND_BE:
-            switch (arenaType)
-            {
-                case ARENA_TYPE_2v2:
-                    return BATTLEGROUND_QUEUE_2v2;
-                case ARENA_TYPE_3v3:
-                    return BATTLEGROUND_QUEUE_3v3;
-                case ARENA_TYPE_5v5:
-                    return BATTLEGROUND_QUEUE_5v5;
-                default:
-                    return BATTLEGROUND_QUEUE_NONE;
-            }
-        default:
-            return BATTLEGROUND_QUEUE_NONE;
-    }
-}
 
-/**
- * @brief Converts a battleground queue type to its template battleground type.
- *
- * Maps queue identifiers back to the battleground template type used to create
- * or reference battleground instances.
- *
- * @param bgQueueTypeId The battleground queue type identifier.
- * @return The corresponding battleground type identifier.
- */
-BattleGroundTypeId BattleGroundMgr::BGTemplateId(BattleGroundQueueTypeId bgQueueTypeId)
-{
-    switch (bgQueueTypeId)
-    {
-        case BATTLEGROUND_QUEUE_WS:
-            return BATTLEGROUND_WS;
-        case BATTLEGROUND_QUEUE_AB:
-            return BATTLEGROUND_AB;
-        case BATTLEGROUND_QUEUE_AV:
-            return BATTLEGROUND_AV;
-        case BATTLEGROUND_QUEUE_EY:
-            return BATTLEGROUND_EY;
-        case BATTLEGROUND_QUEUE_2v2:
-        case BATTLEGROUND_QUEUE_3v3:
-        case BATTLEGROUND_QUEUE_5v5:
-            return BATTLEGROUND_AA;
-        default:
-            return BattleGroundTypeId(0);                   // used for unknown template (it exist and do nothing)
-    }
-}
 
-ArenaType BattleGroundMgr::BGArenaType(BattleGroundQueueTypeId bgQueueTypeId)
-{
-    switch (bgQueueTypeId)
-    {
-        case BATTLEGROUND_QUEUE_2v2:
-            return ARENA_TYPE_2v2;
-        case BATTLEGROUND_QUEUE_3v3:
-            return ARENA_TYPE_3v3;
-        case BATTLEGROUND_QUEUE_5v5:
-            return ARENA_TYPE_5v5;
-        default:
-            return ARENA_TYPE_NONE;
-    }
-}
 
-/**
- * @brief Toggles battleground debug testing mode.
- *
- * Enables or disables testing mode and broadcasts the status change to the world.
- */
-void BattleGroundMgr::ToggleTesting()
-{
-    m_Testing = !m_Testing;
-    if (m_Testing)
-    {
-        sWorld.SendWorldText(LANG_DEBUG_BG_ON);
-    }
-    else
-    {
-        sWorld.SendWorldText(LANG_DEBUG_BG_OFF);
-    }
-}
 
-void BattleGroundMgr::ToggleArenaTesting()
-{
-    m_ArenaTesting = !m_ArenaTesting;
-    if (m_ArenaTesting)
-    {
-        sWorld.SendWorldText(LANG_DEBUG_ARENA_ON);
-    }
-    else
-    {
-        sWorld.SendWorldText(LANG_DEBUG_ARENA_OFF);
-    }
-}
 
-/**
- * @brief Schedules a queue update for a specific battleground queue.
- *
- * Adds a queue update to the scheduler so that the next world update cycle will
- * process matchmaking and invitations for this queue. Multiple requests for the same
- * queue are consolidated to avoid duplicate processing.
- *
- * @param arenaRating
- * @param arenaType
- * @param bgQueueTypeId The battleground queue type to update.
- * @param bgTypeId The battleground type.
- * @param bracket_id The bracket to update.
- */
-void BattleGroundMgr::ScheduleQueueUpdate(uint32 arenaRating, ArenaType arenaType, BattleGroundQueueTypeId bgQueueTypeId, BattleGroundTypeId bgTypeId, BattleGroundBracketId bracket_id)
-{
-    // ACE_Guard<ACE_Thread_Mutex> guard(SchedulerLock);
-    // we will use only 1 number created of bgTypeId and bracket_id
-    uint64 schedule_id = ((uint64)arenaRating << 32) | (arenaType << 24) | (bgQueueTypeId << 16) | (bgTypeId << 8) | bracket_id;
-    bool found = false;
-    for (uint8 i = 0; i < m_QueueUpdateScheduler.size(); ++i)
-    {
-        if (m_QueueUpdateScheduler[i] == schedule_id)
-        {
-            found = true;
-            break;
-        }
-    }
-    if (!found)
-    {
-        m_QueueUpdateScheduler.push_back(schedule_id);
-    }
-}
 
-uint32 BattleGroundMgr::GetMaxRatingDifference() const
-{
-    // this is for stupid people who can't use brain and set max rating difference to 0
-    uint32 diff = sWorld.getConfig(CONFIG_UINT32_ARENA_MAX_RATING_DIFFERENCE);
-    if (diff == 0)
-    {
-        diff = 5000;
-    }
-    return diff;
-}
 
-uint32 BattleGroundMgr::GetRatingDiscardTimer() const
-{
-    return sWorld.getConfig(CONFIG_UINT32_ARENA_RATING_DISCARD_TIMER);
-}
 
-/**
- * @brief Gets the premature finish timer duration.
- *
- * Returns the configured duration in milliseconds after which a battleground can be
- * finished prematurely if one team is significantly outnumbered or defeated.
- *
- * @return The premature finish timer duration in milliseconds.
- */
-uint32 BattleGroundMgr::GetPrematureFinishTime() const
-{
-    return sWorld.getConfig(CONFIG_UINT32_BATTLEGROUND_PREMATURE_FINISH_TIMER);
-}
 
-/**
- * @brief Loads battle master creature entries from the database.
- *
- * Populates the battle master map from the `battlemaster_entry` database table,
- * which maps creature entries to their respective battleground types.
- */
-void BattleGroundMgr::LoadBattleMastersEntry()
-{
-    mBattleMastersMap.clear();                              // need for reload case
 
-    QueryResult* result = WorldDatabase.Query("SELECT `entry`,`bg_template` FROM `battlemaster_entry`");
 
-    uint32 count = 0;
 
-    if (!result)
-    {
-        BarGoLink bar(1);
-        bar.step();
-        sLog.outString(">> Loaded 0 battlemaster entries - table is empty!");
-        sLog.outString();
-        return;
-    }
 
-    BarGoLink bar(result->GetRowCount());
-
-    do
-    {
-        ++count;
-        bar.step();
-
-        Field* fields = result->Fetch();
-
-        uint32 entry = fields[0].GetUInt32();
-        uint32 bgTypeId  = fields[1].GetUInt32();
-        if (!sBattlemasterListStore.LookupEntry(bgTypeId))
-        {
-            sLog.outErrorDb("Table `battlemaster_entry` contain entry %u for nonexistent battleground type %u, ignored.", entry, bgTypeId);
-            continue;
-        }
-
-        mBattleMastersMap[entry] = BattleGroundTypeId(bgTypeId);
-    }
-    while (result->NextRow());
-
-    delete result;
-
-    sLog.outString(">> Loaded %u battlemaster entries", count);
-    sLog.outString();
-}
-
-/**
- * @brief Converts a battleground type to its weekend holiday ID.
- *
- * Maps battleground types to their associated "Call to Arms" weekend holiday events that
- * provide bonus rewards for participating in that battleground type.
- *
- * @param bgTypeId The battleground type to convert.
- * @return The corresponding holiday ID, or HOLIDAY_NONE if not a recognized type.
- */
-HolidayIds BattleGroundMgr::BGTypeToWeekendHolidayId(BattleGroundTypeId bgTypeId)
-{
-    switch (bgTypeId)
-    {
-        case BATTLEGROUND_AV: return HOLIDAY_CALL_TO_ARMS_AV;
-        case BATTLEGROUND_EY: return HOLIDAY_CALL_TO_ARMS_EY;
-        case BATTLEGROUND_WS: return HOLIDAY_CALL_TO_ARMS_WS;
-        case BATTLEGROUND_AB: return HOLIDAY_CALL_TO_ARMS_AB;
-        default: return HOLIDAY_NONE;
-    }
-}
-
-/**
- * @brief Converts a battleground type to its weekend holiday ID.
- *
- * Maps battleground types to their associated "Call to Arms" weekend holiday events.
- *
- * @param holiday The holiday ID to convert.
- * @return The corresponding battleground type, or BATTLEGROUND_TYPE_NONE if invalid.
- */
-BattleGroundTypeId BattleGroundMgr::WeekendHolidayIdToBGType(HolidayIds holiday)
-{
-    switch (holiday)
-    {
-        case HOLIDAY_CALL_TO_ARMS_AV: return BATTLEGROUND_AV;
-        case HOLIDAY_CALL_TO_ARMS_EY: return BATTLEGROUND_EY;
-        case HOLIDAY_CALL_TO_ARMS_WS: return BATTLEGROUND_WS;
-        case HOLIDAY_CALL_TO_ARMS_AB: return BATTLEGROUND_AB;
-        default: return BATTLEGROUND_TYPE_NONE;
-    }
-}
-
-/**
- * @brief Checks if a battleground type is active for the weekend.
- *
- * Determines whether the specified battleground type has an active "Call to Arms"
- * weekend event that provides bonus experience and reputation.
- *
- * @param bgTypeId The battleground type to check.
- * @return true if the battleground is currently featured for the weekend, false otherwise.
- */
-bool BattleGroundMgr::IsBGWeekend(BattleGroundTypeId bgTypeId)
-{
-    return sGameEventMgr.IsActiveHoliday(BGTypeToWeekendHolidayId(bgTypeId));
-}
-
-/**
- * @brief Loads battleground event indexes from the database.
- *
- * Populates the game object and creature event index maps from the database,
- * associating spawned objects and creatures with their battleground events.
- * This enables proper spawning and despawning of objectives during battles.
- */
-void BattleGroundMgr::LoadBattleEventIndexes()
-{
-    BattleGroundEventIdx events;
-    events.event1 = BG_EVENT_NONE;
-    events.event2 = BG_EVENT_NONE;
-    m_GameObjectBattleEventIndexMap.clear();             // need for reload case
-    m_GameObjectBattleEventIndexMap[-1] = events;
-    m_CreatureBattleEventIndexMap.clear();               // need for reload case
-    m_CreatureBattleEventIndexMap[-1] = events;
-
-    uint32 count = 0;
-
-    QueryResult* result =
-    //                                      0             1               2                      3                        4              5                      6
-        WorldDatabase.Query("SELECT `data`.`typ`, `data`.`guid1`, `data`.`ev1` AS `ev1`, `data`.`ev2` AS ev2, `data`.`map` AS m, `data`.`guid2`, `description`.`map`, "
-    //                  7                       8                       9
-        "`description`.`event1`, `description`.`event2`, `description`.`description` "
-        "FROM "
-        "(SELECT '1' AS typ, `a`.`guid` AS `guid1`, `a`.`event1` AS ev1, `a`.`event2` AS ev2, `b`.`map` AS map, `b`.`guid` AS guid2 "
-        "FROM `gameobject_battleground` AS a "
-        "LEFT OUTER JOIN `gameobject` AS b ON `a`.`guid` = `b`.`guid` "
-        "UNION "
-        "SELECT '2' AS typ, `a`.`guid` AS guid1, `a`.`event1` AS ev1, `a`.`event2` AS ev2, `b`.`map` AS map, `b`.`guid` AS guid2 "
-        "FROM `creature_battleground` AS a "
-        "LEFT OUTER JOIN `creature` AS b ON `a`.`guid` = `b`.`guid` "
-        ") data "
-        "RIGHT OUTER JOIN `battleground_events` AS `description` ON `data`.`map` = `description`.`map` "
-        "AND `data`.`ev1` = `description`.`event1` AND `data`.`ev2` = `description`.`event2` "
-    //  full outer join doesn't work in mysql :-/ so just UNION-select the same again and add a left outer join
-        "UNION "
-        "SELECT `data`.`typ`, `data`.`guid1`, `data`.`ev1`, `data`.`ev2`, `data`.`map`, `data`.`guid2`, `description`.`map`, "
-        "`description`.`event1`, `description`.`event2`, `description`.`description` "
-        "FROM "
-        "(SELECT '1' AS typ, `a`.`guid` AS guid1, `a`.`event1` AS ev1, `a`.`event2` AS ev2, `b`.`map` AS map, `b`.`guid` AS guid2 "
-        "FROM `gameobject_battleground` AS a "
-        "LEFT OUTER JOIN `gameobject` AS b ON `a`.`guid` = `b`.`guid` "
-        "UNION "
-        "SELECT '2' AS typ, `a`.`guid` AS guid1, `a`.`event1` AS ev1, `a`.`event2` AS ev2, `b`.`map` AS map, `b`.`guid` AS guid2 "
-        "FROM `creature_battleground` AS a "
-        "LEFT OUTER JOIN `creature` AS b ON `a`.`guid` = `b`.`guid` "
-        ") data "
-        "LEFT OUTER JOIN `battleground_events` AS `description` ON `data`.`map` = `description`.`map` "
-        "AND `data`.`ev1` = `description`.`event1` AND `data`.`ev2` = `description`.`event2` "
-        "ORDER BY `m`, `ev1`, `ev2`");
-    if (!result)
-    {
-        BarGoLink bar(1);
-        bar.step();
-        sLog.outErrorDb(">> Loaded 0 battleground eventindexes.");
-        sLog.outString();
-        return;
-    }
-
-    BarGoLink bar(result->GetRowCount());
-
-    do
-    {
-        bar.step();
-        Field* fields = result->Fetch();
-        if (fields[2].GetUInt8() == BG_EVENT_NONE || fields[3].GetUInt8() == BG_EVENT_NONE)
-        {
-            continue; // we don't need to add those to the eventmap
-        }
-
-        bool gameobject         = (fields[0].GetUInt8() == 1);
-        uint32 dbTableGuidLow   = fields[1].GetUInt32();
-        events.event1           = fields[2].GetUInt8();
-        events.event2           = fields[3].GetUInt8();
-        uint32 map              = fields[4].GetUInt32();
-
-        uint32 desc_map = fields[6].GetUInt32();
-        uint8 desc_event1 = fields[7].GetUInt8();
-        uint8 desc_event2 = fields[8].GetUInt8();
-        const char* description = fields[9].GetString();
-
-        // checking for NULL - through right outer join this will mean following:
-        if (fields[5].GetUInt32() != dbTableGuidLow)
-        {
-            sLog.outErrorDb("BattleGroundEvent: %s with nonexistent guid %u for event: map:%u, event1:%u, event2:%u (\"%s\")",
-                (gameobject) ? "gameobject" : "creature", dbTableGuidLow, map, events.event1, events.event2, description);
-            continue;
-        }
-
-        // checking for NULL - through full outer join this can mean 2 things:
-        if (desc_map != map)
-        {
-            // there is an event missing
-            if (dbTableGuidLow == 0)
-            {
-                sLog.outErrorDb("BattleGroundEvent: missing db-data for map:%u, event1:%u, event2:%u (\"%s\")", desc_map, desc_event1, desc_event2, description);
-                continue;
-            }
-            // we have an event which shouldn't exist
-            else
-            {
-                sLog.outErrorDb("BattleGroundEvent: %s with guid %u is registered, for a nonexistent event: map:%u, event1:%u, event2:%u",
-                    (gameobject) ? "gameobject" : "creature", dbTableGuidLow, map, events.event1, events.event2);
-                continue;
-            }
-        }
-
-        if (gameobject)
-        {
-            m_GameObjectBattleEventIndexMap[dbTableGuidLow] = events;
-        }
-        else
-        {
-            m_CreatureBattleEventIndexMap[dbTableGuidLow] = events;
-        }
-
-        ++count;
-    }
-    while (result->NextRow());
-
-    sLog.outString(">> Loaded %u battleground eventindexes", count);
-    sLog.outString();
-    delete result;
-}
